+++ config/error.js
@@ -1,5 +1,3 @@
-// config/error.js
-
 export class BaseError extends Error {
     constructor(data){
         super(data.message);
