+++ src/models/post.dao.js
@@ -1,0 +1,20 @@
+import { pool } from "../../config/db.config";
+import { BaseError } from "../../config/error";
+import { status } from "../../config/response.status";
+
+//
+export const a = async (b) => {
+    try{
+        const conn = await pool.getConnection();
+        
+        const [confirm] = await pool.query(confirmEmail, b.email);
+
+        const result = await pool.query(insertUserSql, [data.email, data.name, data.gender, data.birth, data.addr, data.specAddr, data.phone]);
+
+        conn.release();
+        return result[0].insertId;
+        
+    }catch (err) {
+        throw new BaseError(status.PARAMETER_IS_WRONG);
+    }
+}
